Partially implemented engine for texture mapping

+++ mandelbulber2/src/ssao_worker.cpp
@@ -124,7 +124,7 @@ void cSSAOWorker::doWork()
 
 				double ambient = 0;
 				double angleStep = M_PI * 2.0 / (double)quality;
-				double maxRandom = 62831 / quality;
+				int maxRandom = 62831 / quality;
 				double rRandom = 1.0;
 
 				if (params->SSAO_random_mode)
